+++ src/main/java/cn/edu/sdju/soft/community/dto/QuestionDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;//和User对象的id对应
+    private Long creator;//和User对象的id对应
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
